# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TeamFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/performance/{period}:
    get:
      operationId: getUserPerformance
      tags: [users]
      summary: Get performance stats for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: period
          in: path
          required: true
          schema:
            type: string
            enum: [day, week, month, year]
      responses:
        "200":
          description: Performance stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPerformance"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (filtered by project/user)
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, overdue]
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (team leader only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{taskId}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as complete (member or system)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # PERFORMANCE
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, healthPoints, performanceScore]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [leader, member]
        healthPoints:
          type: integer
          description: "0-100 health score, decreases with late tasks"
        performanceScore:
          type: number
          description: "0-100 performance percentage"
        avatarUrl:
          type: ["string", "null"]
        githubUsername:
          type: ["string", "null"]
        createdAt:
          type: string
    UserPerformance:
      type: object
      required: [userId, name, performanceScore, healthPoints, tasksCompleted, tasksOnTime, tasksLate, tasksPending, workloadPercentage]
      properties:
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        performanceScore:
          type: number
        healthPoints:
          type: integer
        tasksCompleted:
          type: integer
        tasksOnTime:
          type: integer
        tasksLate:
          type: integer
        tasksPending:
          type: integer
        workloadPercentage:
          type: number
          description: "Percentage of total team work assigned to this user"
...
    Task:
      type: object
      required: [id, title, type, status, projectId, assigneeId, workloadPct, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [programming, documentation, research]
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        assigneeId:
          type: integer
        assigneeName:
          type: ["string", "null"]
        assigneeAvatar:
          type: ["string", "null"]
        workloadPct:
          type: number
          description: "Percentage of work assigned relative to total team workload"
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        githubCommitSha:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, type, projectId, assigneeId, workloadPct]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [programming, documentation, research]
        projectId:
          type: integer
        assigneeId:
          type: integer
        workloadPct:
          type: number
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [programming, documentation, research]
        status:
          type: string
          enum: [pending, in_progress, completed, overdue]
        workloadPct:
          type: number
        dueDate:
          type: string
...