# lib/api-zod/src/generated/types/userPerformance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeamFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPerformance {
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  performanceScore: number;
  healthPoints: number;
  tasksCompleted: number;
  tasksOnTime: number;
  tasksLate: number;
  tasksPending: number;
  /** Percentage of total team work assigned to this user */
  workloadPercentage: number;
}